refactor: rename app/commands to operations
