+++ src/app/mod.rs
@@ -1,5 +1,5 @@
-mod commands;
-pub use commands::*;
+mod operations;
+pub use operations::*;
 
 mod token;
