+++ src/app/operations/user_logout.rs
